+++ CHANGELOG.md
@@ -16,7 +16,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Call method (#55)
   - `services::call`
   - `Node::callMethod`
-- Get/set wrapper types from `Variant` without copy (`Variant::getScalar`, `Variant::getArray`, `Variant::setScalar`, `Variant::setArray`)
+- `Variant`:
+  - Get/set wrapper types without copy (`Variant::getScalar`, `Variant::getArray`, `Variant::setScalar`, `Variant::setArray`)
+  - `const` version of `Variant::getScalar`, `Variant::getArray`
 - Examples
   - `server_method` (#55)
   - `client_method` (#55)